Writing unit tests for the package.
Writing tests for the package. Already wrote tests for the
polynomial related functions. This really is useful, found
several bugs (mostly in the Chebyshev polynomials). I still have to
write tests for the quadrature functions.

One of the bugs I found was a NaN on dUn(x)/dx at -1 and 1. I
rewrote the function using a recurrence relation that I developed.