Corrected bugs in jac_poly.jl that limited precision
When doing calculations using BigFloat, the precision was limited
to Float64 becaus of a bug in several routines. In general, when
calculating weights and zeros, Gamma functions are used and often
a term such as gamma(Q) where Q is an integer is present. In this
case the function returns a Float64, even when calculating BigFloat.

The solution was to replace this term wih something like gamma(Q*one(T))
where T is the type used in the calculations. Now, apparently I can achieve
much greater precisions, compatible with the number type used.