Now the functions work for complex numbers
In several places in gauss_quad.jl, I replaced
AbstractFloat with Number. The functions now work with complex
numbers! Now, I do not know if this is the best strategy. Perhaps
it would be better to keep the functions working for Real numbers only.
For now I will keep it this way.